Cleanup unnecessary files and organize test suite
This change removes a significant number of unnecessary files from the repository, including logs, cache files, backup files, and old test scripts. It also organizes the test suite by separating Jest unit tests from Playwright E2E tests and configures the respective runners to execute them correctly.

- Deleted temporary files: *.log, *.cache, *.bak, error_screenshots.
- Deleted test artifacts: playwright-report/, test-results/.
- Moved unit tests from `public/` to a new `tests/unit/` directory.
- Configured Jest to only run tests from `tests/unit/`.
- Configured Playwright to ignore the `tests/unit/` directory.
- Updated `.gitignore` to prevent temporary files and test artifacts from being committed in the future.
