--- hosting..cache
@@ -1,8 +1,0 @@
-style.css,1752329136199,d7c0c99428068cb0735c3f7846893fb6fa0cbaca940cff6682a1e9b92a943186
-favicon.ico,1752330242171,f61f27bd17de546264aa58f40f3aafaac7021e0ef69c17f6b1b4cd7664a037ec
-README.md,1752773570943,02c3e7fa8af30e68f81e259c94a10f5cb967d992beb7a52c1b4b67be7903a9a4
-.vscode/settings.json,1752328767816,e3b4cefba9302ad06ff24228e50113a846b7a97fb6fa0ebe37b37f46dbf3a171
-.idx/dev.nix,1752328690858,f54580d1309e21f99803f4a61b0c27acf2df71c6b3ab6c60aab3adb8496a3b28
-index.html,1753111525005,4955ef4b3e3c2abc202e1ab43d4bddf21bc9858ee9f221c2ceed3a24154e6d6f
-main.js,1753114783457,6984b54e2e4d0151755b623176b2187ebd342fa7cf00025afe86b4d4c0d6f19f
-firestore.rules,1753111352212,d849a35e3b8ed47d81f855c7d45eaca3873e7684f7a089dec9391511a5807005
