+++ jest.config.js
@@ -3,5 +3,5 @@ export default {
     '^https://www.gstatic.com/firebasejs/9.15.0/firebase-firestore.js$': '<rootDir>/__mocks__/firebase-firestore.js',
   },
   testEnvironment: 'node',
-  testPathIgnorePatterns: ['/node_modules/', '/tests/'],
+  testMatch: ['<rootDir>/tests/unit/**/*.spec.js'],
 };
