+++ package.json
@@ -22,6 +22,7 @@
     "assert": "^2.1.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^30.0.5",
+    "jest-playwright-preset": "^4.0.0",
     "playwright": "^1.55.0",
     "tailwindcss": "^4.1.12"
   },
