+++ playwright.config.js
@@ -1,0 +1,6 @@
+import { defineConfig } from '@playwright/test';
+
+export default defineConfig({
+  testDir: './tests',
+  testIgnore: '**/unit/**',
+});
