--- test-results/.last-run.json
@@ -1,12 +1,0 @@
-{
-  "status": "failed",
-  "failedTests": [
-    "94ab10de85905212603b-477246d3d6fba40dbe72",
-    "94ab10de85905212603b-8268ec9d8138d414a79d",
-    "94ab10de85905212603b-63e93e9a378b99ad6417",
-    "375618813aa2c8aa8a28-f0b6c7c17348f6f4fbb3",
-    "f159fb29fd0c3cefabec-25dc9955245429d8a1b8",
-    "f159fb29fd0c3cefabec-b5ba8b73a3d56b551a08",
-    "f159fb29fd0c3cefabec-b124619cf4e4c15722eb"
-  ]
-}
